Componentise the file browser's path bar
Moves the path breadcrumb markup into its own component along with the
patParts computed property it needs. This is necessary so we can add the
same functionality to the file editor's title.

While doing that I noticed there's a rogue 'set-path' event too where
other places use 'cd', so that's been updated as well.
